feat(contracts): bulk claim rewards (#52)
* feat(contracts): bulk claim rewards

* replay sound

* fix balance

* fix

* sort by total tokens

* fix click

* more fixes

* fix build

* Add claim on starknet

* update correct nums

* add claimed on starknet

+++ contracts/starknet/dojo/src/interfaces/token.cairo
@@ -11,5 +11,5 @@ pub trait IToken<TState> {
 #[starknet::interface]
 pub trait INumsToken<TState> {
     // amount is in 10*18 units
-    fn reward(ref self: TState, recipient: ContractAddress, amount: u32) -> bool;
+    fn reward(ref self: TState, recipient: ContractAddress, amount: u64) -> bool;
 }
